Updated Registration/ Login Page
Updated registration/login page with html and css. Created
RegistrationDOA and User. Reworked RegistrationController.
