+++ web_project/2dv512-app/WebContent/resources/css/index.css
@@ -16,25 +16,44 @@
 .box {
 	position: relative;
 	margin-top: 20px;
-	text-align: center;
 	background: #F8F8F8;
 	font-family: Cabin;
 	-webkit-box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.39);
 	-moz-box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.39);
 	box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.39);
-	padding: 20px;
 	height: 300px;
+	
 }
 
-.box  .login {
-	padding-top: 70px;
+h3 {
+	padding-top: 5px;
+	text-align: center;
 }
 
-.box  .register {
+.box .login {
+	padding-top: 50px;
+	text-align: center;
+}
+
+.box .register {
+	text-align: center;
 	padding-top: 50px;
 }
 
-#error-msg {
-	color: red; padding-left; 5
-	px:
+.box .registerform {
+	padding-top: 20px;
+	padding-left: 20px;
+}
+
+.box .registerform .error {
+	color: red;
+	padding-left: 20px;
+	display: inline;
+}
+.box .complete {
+	text-align: center;
+	padding-top: 70px;
+}
+.inline {
+	display: inline;
 }
