+++ web_project/2dv512-app/src/dv512/model/User.java
@@ -12,8 +12,11 @@ public class User implements Serializable {
 	private static final long serialVersionUID = 8626805850252558885L;
 	
 	private int id = -1;
+	
+	private String name;
+	private String password;
 	private String email;
-
+	
 	public int getId() {
 		return id;
 	}
@@ -21,14 +24,28 @@ public class User implements Serializable {
 	public void setId(int id) {
 		this.id = id;
 	}
-	
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
 
 	public String getEmail() {
 		return email;
 	}
 	
-	public void setEmail(String email) {
-		this.email = email;
+	public String getName() {
+		return name;
+	}
+
+	public String getPassword() {
+		return password;
 	}
 	
 }
